Add files via upload
This app is a dog walker app. It is intended for dog owners who are unable to walk their own dog or may just not have the time that day. It gives owners an option of a female or male walker, how long they would like to have their dog walked (30 or 60 minutes) and if they would like to add a 3 dollar tip to their total purchase. It also asks for users to input their dogs name at the beginning.